fix: correct some comments
