+++ client/public/locales/ru/translation.json
@@ -15,5 +15,8 @@
   "ip": "IP",
   "browser_name": "Название браузера",
   "browser_version": "Версия браузера",
-  "os": "OC"
+  "os": "OC",
+  "en": "Английский",
+  "ru": "Русский",
+  "shorten_url": "Сократить Url!"
 }
