+++ client/src/components/ChangeLanguage/ChangeLanguage.module.scss
@@ -1,0 +1,7 @@
+.choose_wrapper {
+  min-width: 100px;
+}
+
+.choose_list {
+  background-color: #fff;
+}
