+++ client/src/components/Header/Header.module.scss
@@ -11,3 +11,8 @@
     color: #000;
   }
 }
+
+.topBar_wrapper {
+  display: flex;
+  justify-content: space-between;
+}
