+++ client/src/components/ShortnerBtn/ShortnerBtn.tsx
@@ -1,7 +1,8 @@
 import { Box, Button, Collapse } from "@mui/material";
-import { useState } from "react";
+import { useContext, useState } from "react";
 import useActions from "../../hooks/useActions";
 import useTypedSelector from "../../hooks/useTypedSelector";
+import { CONTEXT } from "../../libs/constants";
 import { IShortnerBtnProps } from "../../types/types";
 import Loader from "../Loader/Loader";
 import ShortUrl from "../ShortUrl/ShortUrl";
@@ -10,6 +11,7 @@ import styles from "./ShortnerBtn.module.scss";
 
 function ShortnerBtn(props: IShortnerBtnProps) {
   const { url, isError } = props;
+  const { translation } = useContext(CONTEXT);
   const { link, error, loading } = useTypedSelector((state) => state.link);
   const [visible, setVisible] = useState(false);
   const { fetchLink } = useActions();
@@ -28,7 +30,7 @@ function ShortnerBtn(props: IShortnerBtnProps) {
         onClick={cutUrl}
         className={styles.btn__cut}
       >
-        Shorten Url!
+        {translation("shorten_url")}
       </Button>
       <Collapse in={visible}>
         {loading ? (
