+++ server/package.json
@@ -46,6 +46,7 @@
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "geoip-lite": "^1.4.7",
+    "http-status-codes": "^2.2.0",
     "ip": "^1.1.8",
     "pg": "^8.10.0",
     "pg-hstore": "^2.3.4",
