+++ server/src/error/apiError.ts
@@ -8,18 +8,6 @@ class ApiError extends Error {
     this.status = status;
     this.message = message;
   }
-
-  static badRequest(message: string) {
-    return new ApiError(404, message);
-  }
-
-  static internal(message: string) {
-    return new ApiError(500, message);
-  }
-
-  static forbidden(message: string) {
-    return new ApiError(403, message);
-  }
 }
 
 export default ApiError;
